university_apis/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
dj-database-url==0.5.0
Django==3.0.7
django-countries==6.1.2
django-elasticsearch-dsl==7.1.4
django-elasticsearch-dsl-drf==0.20.8
django-filter==2.3.0
django-heroku==0.3.1
django-nine==0.2.3
djangorestframework==3.11.0
elasticsearch==7.8.0
elasticsearch-dsl==7.2.1
gunicorn==20.0.4
idna==2.10
mysqlclient==1.4.6
psycopg2==2.8.5
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
validators==0.16.0
whitenoise==5.1.0
